Directory refactoring + cleaning
Trying to make the structure more readable by adhering to the advice found on the async_redux page. https://pub.dev/packages/async_redux#recommended-directory-structure
