--- client/lib/containers/app_loading.dart
@@ -1,21 +1,0 @@
-//import 'package:betterstatmobile_business_logic/actions/actions.dart';
-//import 'package:betterstatmobile_business_logic/models/models.dart';
-//import 'package:betterstatmobile_client_components/containers/typedefs.dart';
-//import 'package:flutter/foundation.dart';
-//import 'package:flutter/widgets.dart';
-//
-//class AppLoading extends StoreConnector<AppState, AppActions, bool> {
-//  final ViewModelBuilder<bool> builder;
-//
-//  AppLoading({Key key, @required this.builder}) : super(key: key);
-//
-//  @override
-//  Widget build(BuildContext context, bool state, AppActions actions) {
-//    return builder(context, state);
-//  }
-//
-//  @override
-//  bool connect(AppState state) {
-//    return state.isLoading;
-//  }
-//}
