+++ client/lib/day/connector/add_day.dart
